Move tests, start integrating introspection types
