+++ lib/absinthe/type/built_ins.ex
@@ -4,5 +4,6 @@ defmodule Absinthe.Type.BuiltIns do
 
   import_types BuiltIns.Scalars
   import_types BuiltIns.Directives
+  import_types BuiltIns.Introspection
 
 end
