--- test/specification/execution/grouped_field_set_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Execution.GroupedFieldSetTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Evaluating-a-grouped-field-set"
-
-end
