--- test/specification/execution/operations_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Execution.OperationsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Evaluating-operations"
-
-end
