--- test/specification/execution/requests_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Execution.RequestsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Evaluting-requests"
-
-end
