--- test/specification/execution/selection_sets_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Execution.SelectionSetsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Evaluating-selection-sets"
-
-end
