--- test/specification/execution/variables_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Execution.VariablesTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Coercing-variables"
-
-end
