--- test/specification/validation/arguments_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.ArgumentsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Validation.Arguments"
-
-end
