--- test/specification/validation/directives_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.DirectivesTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Validation.Directives"
-
-end
