--- test/specification/validation/fields_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.FieldsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Validation.Fields"
-
-end
