--- test/specification/validation/fragments_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.FragmentsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Validation.Fragments"
-
-end
