--- test/specification/validation/operations_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.OperationsTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Validation.Operations"
-
-end
