--- test/specification/validation/values_test.exs
@@ -1,7 +1,0 @@
-defmodule Specification.Validation.ValuesTest do
-  use ExSpec, async: true
-  @moduletag specification: true, pending: true
-
-  @graphql_spec "#sec-Values"
-
-end
