+++ test/support/support_schemas.ex
@@ -22,14 +22,15 @@ defmodule SupportSchemas do
     err = assert_raise Absinthe.Schema.Error, fn ->
       load_schema(schema_name)
     end
-    patterns
-    |> Enum.each(fn
+    found = patterns
+    |> Enum.filter(fn
       pattern ->
         assert Enum.find(err.details, fn
           detail ->
             pattern.rule == detail.rule && pattern.data == detail.data
-        end)
+        end), "Could not find error detail pattern #{inspect pattern} in #{inspect err.details}"
     end)
+    assert length(patterns) == length(err.details)
   end
 
 end
